HW-2 retrofit api interface division object and MainActivity class retrofit api call code update
